Move TimeAgo to language
