+++ go/meshwrapper/helpers/language.go
@@ -1,5 +1,11 @@
 package helpers
 
+import (
+	"fmt"
+	"math"
+	"time"
+)
+
 func Pluralize(word string, count int) string {
 	if count == 1 {
 		return word
@@ -8,4 +14,37 @@ func Pluralize(word string, count int) string {
 		return "them"
 	}
 	return word + "s"
+}
+
+func TimeAgo(timestamp time.Time) string {
+	seconds := int(time.Since(timestamp).Seconds())
+
+	if seconds == 1 {
+		return "one second"
+	}
+	if seconds < 60 {
+		return fmt.Sprintf("%d seconds", seconds)
+	}
+
+	minutes := int(math.Floor(float64(seconds) / 60))
+	if minutes == 1 {
+		return "one minute"
+	}
+	if minutes < 60 {
+		return fmt.Sprintf("%d minutes", minutes)
+	}
+
+	hours := int(math.Floor(float64(minutes) / 60))
+	if hours == 1 {
+		return "one hour"
+	}
+	if hours < 24 {
+		return fmt.Sprintf("%d hours", hours)
+	}
+
+	days := int(math.Floor(float64(hours) / 24))
+	if days == 1 {
+		return "one day"
+	}
+	return fmt.Sprintf("%d days", days)
 }
